requirements.txt
langchain==0.1.12
langchain-openai==0.0.8
openai==1.14.0
python-dotenv==1.0.1
streamlit==1.32.2
typing-extensions>=4.5.0
pydantic>=2.0.0
aiohttp>=3.8.0
tenacity>=8.0.0


